refactor styling and convention form folder
