+++ src/components/form/Checkbox/Checkbox.style.ts
@@ -1,16 +1,18 @@
-import { scale, styleManager } from '@/libs'
+import { styleManager } from '@/libs'
 
-export const checkboxStyles = styleManager.createStyleSheet(theme => ({
-  wrapper: {
-    width: scale(24),
-    height: scale(24),
-    borderWidth: scale(1),
-    borderRadius: 2,
-  },
-  wrapperReadOnly: {
-    opacity: 0.4,
-  },
-  wrapperChecked: {
-    backgroundColor: theme.colors.blue,
-  },
-}))
+export const stylesheet = styleManager.createStyleSheet(
+  ({ colors, utils: { scale } }) => ({
+    wrapper: {
+      width: scale(24),
+      height: scale(24),
+      borderWidth: scale(1),
+      borderRadius: 2,
+    },
+    wrapperReadOnly: {
+      opacity: 0.4,
+    },
+    wrapperChecked: {
+      backgroundColor: colors.blue,
+    },
+  }),
+)
