+++ src/components/form/Input/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Input'
