+++ src/components/form/Radio/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Radio'
