+++ src/components/form/Switch/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Switch'
