+++ src/components/form/TextArea/index.tsx
@@ -1,58 +1,1 @@
-import React, { FC } from 'react'
-import { Controller, useFormContext } from 'react-hook-form'
-import { TextInput, TextInputProps } from 'react-native'
-
-import { Box, ErrorText } from '@/components'
-import { useDisclose } from '@/hooks'
-import { styleManager } from '@/libs'
-
-import { textAreaStyles } from './TextArea.style'
-
-type InputProps = {
-  name: string
-  readOnly?: boolean
-} & TextInputProps
-
-const TextArea: FC<InputProps> = ({ name, style, ...rest }) => {
-  const { isOpen: isFocused, open: onFocus, close: onBlur } = useDisclose()
-  const {
-    control,
-    formState: { errors },
-  } = useFormContext()
-  const { theme, styles } = styleManager.useStyles(textAreaStyles)
-  const hasError = errors[name] ? true : false
-
-  return (
-    <Controller
-      control={control}
-      name={name}
-      render={({ field }) => (
-        <Box>
-          <TextInput
-            value={field.value && field.value.toString()}
-            onChangeText={field.onChange}
-            onBlur={() => {
-              field.onBlur()
-              onBlur()
-            }}
-            onFocus={onFocus}
-            placeholderTextColor={theme.colors.gray}
-            style={[
-              styles.textArea,
-              rest.readOnly && styles.readOnlyInput,
-              isFocused && styles.focusedInput,
-              hasError && styles.errorInput,
-              style,
-            ]}
-            textAlignVertical="top"
-            multiline
-            {...rest}
-          />
-          <ErrorText name={name} errors={errors} />
-        </Box>
-      )}
-    />
-  )
-}
-
-export default TextArea
+export { default } from './TextArea'
